# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BabyAI API
  version: 0.1.0
  description: >
    BabyAI MVP API. 육아 기록, AI 조회, 리포트, 구독/사진 공유, Siri/Bixby 연동을 포함한다.
servers:
  - url: http://localhost:8000
    description: local
...
paths:
...
  /api/v1/onboarding/parent:
    post:
      tags: [onboarding]
      summary: Parent onboarding with required baby profile
      operationId: onboardingParent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentOnboardingRequest'
      responses:
        '200':
          description: Parent onboarding completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: created }
                  household_id: { type: string }
                  baby_profile_created: { type: boolean, example: true }
                  provider: { $ref: '#/components/schemas/AuthProvider' }
  /api/v1/events/voice:
    post:
      tags: [events]
      summary: Parse voice recording into structured events
      operationId: parseVoiceEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceUploadRequest'
      responses:
        '200':
          description: Voice parsed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceParseResponse'
  /api/v1/events/confirm:
    post:
      tags: [events]
      summary: Confirm parsed events with one-tap action
      operationId: confirmEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmEventsRequest'
      responses:
        '200':
          description: Confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: CONFIRMED }
                  clip_id: { type: string }
                  saved_event_count: { type: integer }
  /api/v1/events/manual:
    post:
      tags: [events]
      summary: Create a completed manual event in one request
      operationId: createManualEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualEventCreateRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: CREATED }
                  event_id: { type: string }
                  type: { type: string }
  /api/v1/events/start:
    post:
      tags: [events]
      summary: Start an in-progress event (OPEN state)
      operationId: startManualEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualEventStartRequest'
      responses:
        '200':
          description: Started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: STARTED }
                  event_id: { type: string }
                  type: { type: string }
                  start_time: { type: string, format: date-time }
                  event_state: { type: string, example: OPEN }
  /api/v1/events/{event_id}/complete:
    patch:
      tags: [events]
      summary: Complete an OPEN event
      operationId: completeManualEvent
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualEventCompleteRequest'
      responses:
        '200':
          description: Completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: COMPLETED }
                  event_id: { type: string }
                  type: { type: string }
                  start_time: { type: string, format: date-time }
                  end_time: { type: string, format: date-time }
                  duration_min: { type: integer }
                  event_state: { type: string, example: CLOSED }
  /api/v1/events/{event_id}/cancel:
    patch:
      tags: [events]
      summary: Cancel an OPEN event
      operationId: cancelManualEvent
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualEventCancelRequest'
      responses:
        '200':
          description: Canceled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: CANCELED }
                  event_id: { type: string }
                  type: { type: string }
                  event_state: { type: string, example: CANCELED }
  /api/v1/events/open:
    get:
      tags: [events]
      summary: List OPEN events for a baby
      operationId: listOpenEvents
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [FORMULA, BREASTFEED, SLEEP, PEE, POO, GROWTH, MEMO, SYMPTOM, MEDICATION]
      responses:
        '200':
          description: Open events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenEventsResponse'
  /api/v1/settings/me:
    get:
      tags: [settings]
      summary: Get current app settings for authenticated user
      operationId: getMySettings
      responses:
        '200':
          description: User settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponse'
    patch:
      tags: [settings]
      summary: Update current app settings for authenticated user
      operationId: updateMySettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsUpdateRequest'
      responses:
        '200':
          description: Updated user settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsResponse'
  /api/v1/data/export.csv:
    get:
      tags: [settings]
      summary: Export baby event data as CSV
      operationId: exportBabyDataCsv
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: CSV data export
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/v1/babies/profile:
    get:
      tags: [babies]
      summary: Get baby profile and feeding recommendation
      operationId: getBabyProfile
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Baby profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BabyProfileResponse'
    patch:
      tags: [babies]
      summary: Upsert baby profile settings
      operationId: upsertBabyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BabyProfileUpdateRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BabyProfileResponse'
  /api/v1/ai/query:
    post:
      tags: [ai]
      summary: Ask chatbot with optional personal-data mode
      operationId: aiQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQueryRequest'
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiQueryResponse'
...
  /api/v1/quick/last-poo-time:
    get:
      tags: [ai]
      summary: Get the last poo event time
      operationId: quickLastPooTime
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
        - in: query
          name: tone
          schema: { $ref: '#/components/schemas/AiTone' }
      responses:
        '200':
          description: Last poo time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPooTimeResponse'
  /api/v1/quick/next-feeding-eta:
    get:
      tags: [ai]
      summary: Get next feeding ETA from recent feeding intervals
      operationId: quickNextFeedingEta
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
        - in: query
          name: tone
          schema: { $ref: '#/components/schemas/AiTone' }
      responses:
        '200':
          description: Next feeding ETA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextFeedingEtaResponse'
  /api/v1/quick/today-summary:
    get:
      tags: [ai]
      summary: Get today's summary lines
      operationId: quickTodaySummary
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Today summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodaySummaryResponse'
  /api/v1/quick/landing-snapshot:
    get:
      tags: [ai]
      summary: Get landing dashboard snapshot for day/week/month records
      operationId: quickLandingSnapshot
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
        - in: query
          name: range
          required: false
          schema:
            type: string
            enum: [day, week, month]
        - in: query
          name: tz_offset
          required: false
          schema: { type: string, example: '+09:00' }
      responses:
        '200':
          description: Landing dashboard snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LandingSnapshotResponse'
  /api/v1/reports/daily:
    get:
      tags: [reports]
      summary: Daily report
      operationId: getDailyReport
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Daily report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReportResponse'
  /api/v1/reports/weekly:
    get:
      tags: [reports]
      summary: Weekly report
      operationId: getWeeklyReport
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
        - in: query
          name: week_start
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyReportResponse'
  /api/v1/photos/upload-url:
    post:
      tags: [photos]
      summary: Create presigned upload URL
      operationId: createPhotoUploadUrl
      parameters:
        - in: query
          name: album_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Presigned URL created
          content:
            application/json:
              schema:
                type: object
                properties:
                  album_id: { type: string }
                  upload_url: { type: string, format: uri }
                  object_key: { type: string }
  /api/v1/photos/complete:
    post:
      tags: [photos]
      summary: Confirm photo upload and generate variants metadata
      operationId: completePhotoUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoUploadCompleteRequest'
      responses:
        '200':
          description: Upload completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: uploaded }
                  photo_id: { type: string }
                  album_id: { type: string }
                  downloadable: { type: boolean }
                  variants:
                    type: object
                    additionalProperties: { type: string }
...
  /api/v1/subscription/me:
    get:
      tags: [subscription]
      summary: Get household subscription
      operationId: getMySubscription
      parameters:
        - in: query
          name: household_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Subscription snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  household_id: { type: string }
                  plan: { $ref: '#/components/schemas/SubscriptionPlan' }
                  status: { type: string, example: active }
  /api/v1/subscription/checkout:
    post:
      tags: [subscription]
      summary: Start checkout
      operationId: checkoutSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: pending_payment }
                  plan: { $ref: '#/components/schemas/SubscriptionPlan' }
                  household_id: { type: string }
  /api/v1/assistants/siri/{intentName}:
    post:
      tags: [assistants]
      summary: Siri App Intent dialog endpoint
      operationId: executeSiriIntent
      parameters:
        - in: path
          name: intentName
          required: true
          schema:
            type: string
            enum: [GetLastPooTime, GetNextFeedingEta, GetTodaySummary]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiriIntentRequest'
      responses:
        '200':
          description: Dialog response for Siri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantDialogResponse'
  /api/v1/assistants/bixby/query:
    post:
      tags: [assistants]
      summary: Bixby endpoint query
      operationId: executeBixbyQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BixbyQueryRequest'
      responses:
        '200':
          description: Bixby answer payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  resultMoment: { type: boolean, example: true }
components:
  schemas:
    AuthProvider:
      type: string
      enum: [apple, google, phone]
    AiTone:
      type: string
      enum: [friendly, neutral, formal, brief, coach]
...
    ParentOnboardingRequest:
      type: object
      required: [provider, baby_name, baby_birth_date, required_consents]
      properties:
        provider: { $ref: '#/components/schemas/AuthProvider' }
        baby_name: { type: string }
        baby_birth_date: { type: string, format: date }
        baby_sex:
          type: string
          enum: [male, female, other, unknown]
        baby_weight_kg: { type: number, format: float }
        feeding_method:
          type: string
          enum: [formula, breastmilk, mixed]
        formula_brand: { type: string }
        formula_product: { type: string }
        formula_type: { type: string, example: standard }
        formula_contains_starch: { type: boolean }
        required_consents:
          type: array
          items:
            type: string
            enum: [terms, privacy, data_processing]
    VoiceUploadRequest:
      type: object
      required: [baby_id]
      properties:
        baby_id: { type: string }
        transcript_hint: { type: string }
...
    VoiceParseResponse:
      type: object
      required: [clip_id, transcript, parsed_events, status]
      properties:
        clip_id: { type: string }
        transcript: { type: string }
        parsed_events:
          type: array
          items: { $ref: '#/components/schemas/EventItem' }
        status:
          type: string
          enum: [PARSED, FAILED]
    ConfirmEventsRequest:
      type: object
      required: [clip_id, events]
      properties:
        clip_id: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/EventItem' }
    ManualEventCreateRequest:
      type: object
      required: [baby_id, type, start_time]
      properties:
        baby_id: { type: string }
        type:
          type: string
          enum: [FORMULA, BREASTFEED, SLEEP, PEE, POO, GROWTH, MEMO, SYMPTOM, MEDICATION]
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time, nullable: true }
        value:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    ManualEventStartRequest:
      type: object
      required: [baby_id, type, start_time]
      properties:
        baby_id: { type: string }
        type:
          type: string
          enum: [FORMULA, BREASTFEED, SLEEP, PEE, POO, MEDICATION, MEMO]
        start_time: { type: string, format: date-time }
        value:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    ManualEventCompleteRequest:
      type: object
      properties:
        end_time: { type: string, format: date-time, nullable: true }
        value:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    ManualEventCancelRequest:
      type: object
      properties:
        reason: { type: string }
...
    OpenEventsResponse:
      type: object
      properties:
        baby_id: { type: string }
        open_events:
          type: array
          items: { $ref: '#/components/schemas/OpenEventItem' }
        open_count: { type: integer }
        event_state: { type: string, example: OPEN }
        reference_text: { type: string }
...
    BabyProfileUpdateRequest:
      type: object
      required: [baby_id]
      properties:
        baby_id: { type: string }
        baby_name: { type: string }
        baby_birth_date: { type: string, format: date }
        baby_sex:
          type: string
          enum: [male, female, other, unknown]
        baby_weight_kg: { type: number, format: float }
        feeding_method:
          type: string
          enum: [formula, breastmilk, mixed]
        formula_brand: { type: string }
        formula_product: { type: string }
        formula_type: { type: string }
        formula_contains_starch: { type: boolean }
...